requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Django==2.1
django-crispy-forms==1.9.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
idna==2.9
idna-ssl==1.1.0
jdcal==1.4.1
multidict==4.7.5
mysql-connector-python-rf==2.2.2
numpy==1.18.1
pandas==1.0.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
typing-extensions==3.7.4.2
urllib3==1.25.8
xlrd==1.2.0
yarl==1.4.2
